% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_display_options.R
\name{add,ANY,plot_display_options-method}
\alias{add,ANY,plot_display_options-method}
\title{Apply plot options to plot.}
\usage{
\S4method{add}{ANY,plot_display_options}(object, x, variable)
}
\arguments{
\item{object}{plot}

\item{x}{options}

\item{variable}{variable of interest}
}
\value{
updated plot
}
\description{
Apply plot options to plot.
}
