% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsismap_model.R
\name{add,campsismap_model,error_model-method}
\alias{add,campsismap_model,error_model-method}
\title{Add error model to Campsismap model.}
\usage{
\S4method{add}{campsismap_model,error_model}(object, x)
}
\arguments{
\item{object}{Campsismap model}

\item{x}{error model}
}
\value{
updated Campsismap model
}
\description{
Add error model to Campsismap model.
}
