% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendation.R
\name{add,recommendation,observations-method}
\alias{add,recommendation,observations-method}
\title{Add observations (for simulation) to both the original and recommended datasets.}
\usage{
\S4method{add}{recommendation,observations}(object, x)
}
\arguments{
\item{object}{recommendation}

\item{x}{observations}
}
\value{
updated Campsismap model
}
\description{
Add observations (for simulation) to both the original and recommended datasets.
}
