% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/dataset.R
\name{addDV}
\alias{addDV}
\alias{addDV,dataset,tbl_df-method}
\alias{addDV,dataset,data.frame-method}
\title{Add information related to the dependent variable (DV), i.e. add the observed samples.}
\usage{
addDV(object, x, ...)

\S4method{addDV}{dataset,tbl_df}(object, x)

\S4method{addDV}{dataset,data.frame}(object, x)
}
\arguments{
\item{object}{generic object}

\item{x}{data frame with 2 columns: TIME and DV}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Add information related to the dependent variable (DV), i.e. add the observed samples.
}
