% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/dataset.R
\name{addSamples}
\alias{addSamples}
\alias{addSamples,dataset,tbl_df-method}
\alias{addSamples,dataset,data.frame-method}
\title{Add the observed samples.}
\usage{
addSamples(object, x, ...)

\S4method{addSamples}{dataset,tbl_df}(object, x)

\S4method{addSamples}{dataset,data.frame}(object, x)
}
\arguments{
\item{object}{generic object}

\item{x}{data frame with 2 columns: TIME and DV}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Add the observed samples.
}
