% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.R
\docType{class}
\name{combined_error_model-class}
\alias{combined_error_model-class}
\title{Combined error model class.}
\description{
Combined error model class.
}
\section{Slots}{

\describe{
\item{\code{prop}}{standard deviation of proportional error, numeric}

\item{\code{add}}{additive error, same unit as observed variable, numeric}
}}

