% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.R
\name{CombinedErrorModel}
\alias{CombinedErrorModel}
\title{Combined error model.}
\usage{
CombinedErrorModel(prop, add)
}
\arguments{
\item{prop}{standard deviation of proportional error, numeric}

\item{add}{standard deviation of additive error, same unit as observed variable, numeric}
}
\value{
error model
}
\description{
Combined error model.
}
