% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/error_model.R
\name{computeSd}
\alias{computeSd}
\alias{computeSd,proportional_error_model,numeric-method}
\title{Compute standard deviation of error model.}
\usage{
computeSd(object, ipred, ...)

\S4method{computeSd}{proportional_error_model,numeric}(object, ipred)
}
\arguments{
\item{object}{error model}

\item{ipred}{individual predicted values}

\item{...}{extra arguments}
}
\value{
the standard deviation
}
\description{
Compute standard deviation of error model.
}
