% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_rounding_rule.R
\name{DoseRoundingRule}
\alias{DoseRoundingRule}
\title{Create a dose rounding rule.}
\usage{
DoseRoundingRule(fun = ~round(.x))
}
\arguments{
\item{fun}{purrr-style lambda function used for rounding doses, first argument is the dose amount to round}
}
\description{
Create a dose rounding rule.
}
