% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_dosing_table_editor.R
\name{DosingTableEditor}
\alias{DosingTableEditor}
\title{Dosing table editor.}
\usage{
DosingTableEditor(
  tableReact = NULL,
  ns = shiny::NS("dosing_table"),
  fun = NULL,
  initialDose = 100,
  initialTime = "08:00"
)
}
\arguments{
\item{tableReact}{reactive table}

\item{ns}{shiny namespace}

\item{fun}{conversion function (table -> anything)}

\item{initialDose}{first dose in table, only if tableReact is NULL}

\item{initialTime}{first dose time in table, only if tableReact is NULL}
}
\description{
Dosing table editor.
}
