% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plot.R
\name{drawTarget}
\alias{drawTarget}
\title{Draw target profile.}
\usage{
drawTarget(x, target, colour, linetype = "dashed")
}
\arguments{
\item{x}{ggplot object}

\item{target}{target effective definition, an object of class 'target_definition_effective'}

\item{colour}{colour of the target profile}

\item{linetype}{linetype of the target profile, default is 'dashed'}
}
\value{
updated ggplot object
}
\description{
Draw target profile.
}
