% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.R
\docType{class}
\name{error_model-class}
\alias{error_model-class}
\title{Error model class.}
\description{
Error model class.
}
\section{Slots}{

\describe{
\item{\code{fun}}{function to compute the standard deviation of the model}
}}

