% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/map_estimation.R
\name{estimate}
\alias{estimate}
\alias{estimate,campsismap_model,dataset,numeric-method}
\title{Estimate.}
\usage{
estimate(model, dataset, etas = NULL, ...)

\S4method{estimate}{campsismap_model,dataset,numeric}(model, dataset, etas = NULL, ...)
}
\arguments{
\item{model}{Campsismap model}

\item{dataset}{predicted values, numeric vector}

\item{etas}{initial etas vector, default will be all etas 0}

\item{...}{extra arguments}
}
\value{
estimated etas, numeric vector
}
\description{
Estimate.
}
