% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_datetime_table_editor.R
\name{exportDatetimeTable}
\alias{exportDatetimeTable}
\title{Export datetime table according to given function.}
\usage{
exportDatetimeTable(table, fun, dateTime0)
}
\arguments{
\item{table}{table content (see slot tableReact)}

\item{fun}{any user given function}

\item{dateTime0}{reference datetime}
}
\value{
a list, each element is the result of the function applied to a row of the table
}
\description{
Export datetime table according to given function.
}
