% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_inputs.R
\name{findAll}
\alias{findAll}
\alias{findAll,shiny_inputs,shiny_input-method}
\title{Find all elements. A list object is returned with all requested elements.}
\usage{
findAll(object, x, ...)

\S4method{findAll}{shiny_inputs,shiny_input}(object, x, ...)
}
\arguments{
\item{object}{list object}

\item{x}{first element to find}

\item{...}{other elements to find}
}
\value{
filtered list object
}
\description{
Find all elements. A list object is returned with all requested elements.
}
