% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_inputs.R
\name{findInput}
\alias{findInput}
\title{Find given input in object.}
\usage{
findInput(object, x)
}
\arguments{
\item{object}{list of inputs}

\item{x}{input}
}
\value{
input stored in object
}
\description{
Find given input in object.
}
