% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_generic.R
\name{getDialog}
\alias{getDialog}
\title{Get dialog.}
\usage{
getDialog(object, failed, errorMessage = NULL, code)
}
\arguments{
\item{object}{generic object}

\item{failed}{logical}

\item{errorMessage}{error message}

\item{code}{code to be shown}
}
\value{
a dialog to be shown
}
\description{
Get dialog.
}
