% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/dataset.R
\name{getDV}
\alias{getDV}
\alias{getDV,dataset-method}
\title{Get information related to the dependent variable (DV), i.e. add the observed samples.}
\usage{
getDV(object, ...)

\S4method{getDV}{dataset}(object)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments}
}
\value{
tibble with the samples (TIME and DV)
}
\description{
Get information related to the dependent variable (DV), i.e. add the observed samples.
}
