% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_inputs.R
\name{getEntries}
\alias{getEntries}
\alias{getEntries,shiny_inputs-method}
\title{Get entries.}
\usage{
getEntries(object)

\S4method{getEntries}{shiny_inputs}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
named vector
}
\description{
Get entries.
}
