% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_generic.R, R/shiny_input.R
\name{getID}
\alias{getID}
\alias{getID,shiny_input-method}
\title{Get identifier.}
\usage{
getID(object)

\S4method{getID}{shiny_input}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
identifier, character
}
\description{
Get identifier.
}
