% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_datetime_table_editor.R,
%   R/shiny_dosing_table_editor.R, R/shiny_observations_table_editor.R
\name{getInitialTable}
\alias{getInitialTable}
\alias{getInitialTable,dosing_table_editor-method}
\alias{getInitialTable,observations_table_editor-method}
\title{Get initial table.}
\usage{
getInitialTable(object, ...)

\S4method{getInitialTable}{dosing_table_editor}(object, init_dose, init_time)

\S4method{getInitialTable}{observations_table_editor}(object, init_obs, init_time)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments}
}
\value{
initial table, tibble
}
\description{
Get initial table.
}
