% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_inputs.R
\name{getInputID}
\alias{getInputID}
\alias{getInputID,shiny_inputs,shiny_input-method}
\title{Get input ID (including namespace).}
\usage{
getInputID(object, x)

\S4method{getInputID}{shiny_inputs,shiny_input}(object, x)
}
\arguments{
\item{object}{generic object}

\item{x}{input to get the ID}
}
\value{
input ID, including namespace
}
\description{
Get input ID (including namespace).
}
