% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_inputs.R
\name{getInputValue}
\alias{getInputValue}
\alias{getInputValue,shiny_inputs,shiny_input-method}
\title{Get input value.}
\usage{
getInputValue(object, x)

\S4method{getInputValue}{shiny_inputs,shiny_input}(object, x)
}
\arguments{
\item{object}{generic object}

\item{x}{input to get the ID}
}
\value{
input value
}
\description{
Get input value.
}
