% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_datetime_table_editor.R
\name{getReferenceDateTime}
\alias{getReferenceDateTime}
\title{Get the reference date time.}
\usage{
getReferenceDateTime(table)
}
\arguments{
\item{table}{table content}
}
\description{
Get the reference date time.
}
