% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/dataset.R
\name{getSamples}
\alias{getSamples}
\alias{getSamples,dataset-method}
\title{Get the observed samples.}
\usage{
getSamples(object, ...)

\S4method{getSamples}{dataset}(object)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments}
}
\value{
tibble with the samples (TIME and DV)
}
\description{
Get the observed samples.
}
