% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/dataset.R
\name{getSimulationTimes}
\alias{getSimulationTimes}
\alias{getSimulationTimes,dataset-method}
\title{Get the simulation times (by opposition to the samples).}
\usage{
getSimulationTimes(object, ...)

\S4method{getSimulationTimes}{dataset}(object)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments}
}
\value{
a numeric vector with the simulation times
}
\description{
Get the simulation times (by opposition to the samples).
}
