% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_generic.R,
%   R/shiny_datetime_table_editor.R
\name{getUI}
\alias{getUI}
\alias{getUI,datetime_table_editor-method}
\title{Get user interface.}
\usage{
getUI(object)

\S4method{getUI}{datetime_table_editor}(object)
}
\arguments{
\item{object}{generic object}
}
\value{
user interface (shiny component)
}
\description{
Get user interface.
}
