% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plot_type.R
\name{IndividualFitPlotType}
\alias{IndividualFitPlotType}
\title{Individual fit plot type.}
\usage{
IndividualFitPlotType(show_pop = TRUE)
}
\arguments{
\item{show_pop}{show population (typical profile)}
}
\value{
an individual fit plot type
}
\description{
Individual fit plot type.
}
