% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{individualPrediction}
\alias{individualPrediction}
\title{Individual predictions calculation.}
\usage{
individualPrediction(model, dataset, etas)
}
\arguments{
\item{model}{Campsismap model}

\item{dataset}{Campsis dataset}

\item{etas}{unnamed numeric vector}
}
\value{
a dataframe with 3 columns: TIME, <VARIABLE>, DV
}
\description{
Individual predictions calculation.
}
