% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/model_cache.R
\name{initCache}
\alias{initCache}
\alias{initCache,rxode2_model_cache,campsis_model,simulation_settings-method}
\alias{initCache,mrgsolve_model_cache,campsis_model,simulation_settings-method}
\title{Init the model cache.}
\usage{
initCache(object, model, settings, ...)

\S4method{initCache}{rxode2_model_cache,campsis_model,simulation_settings}(object, model, settings, ...)

\S4method{initCache}{mrgsolve_model_cache,campsis_model,simulation_settings}(object, model, settings, ...)
}
\arguments{
\item{object}{model cache}

\item{model}{original model}

\item{settings}{simulation settings}

\item{...}{extra arguments}
}
\value{
updated model
}
\description{
Init the model cache.
}
