% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{initialiseEtaVector}
\alias{initialiseEtaVector}
\title{Initialise the ETA vector based on the Campsismap model.
ETA vector is initialised only if it has length 0.}
\usage{
initialiseEtaVector(etas, model)
}
\arguments{
\item{etas}{given ETA vector}

\item{model}{Campsismap model}
}
\value{
initiated ETA vector
}
\description{
Initialise the ETA vector based on the Campsismap model.
ETA vector is initialised only if it has length 0.
}
