% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_definition.R
\name{length,target_definition-method}
\alias{length,target_definition-method}
\title{Return the number of rows contained in the target definition.}
\usage{
\S4method{length}{target_definition}(x)
}
\arguments{
\item{x}{target definition}
}
\value{
a number
}
\description{
Return the number of rows contained in the target definition.
}
