% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_options.R
\name{maxValueInPlot}
\alias{maxValueInPlot}
\title{Automatically get the max value from the plot.}
\usage{
maxValueInPlot(plot, variable)
}
\arguments{
\item{plot}{ggplot2 plot}

\item{variable}{variable of interest}
}
\value{
max value in all data layers
}
\description{
Automatically get the max value from the plot.
}
