% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_options.R
\name{minorBreaksCustom}
\alias{minorBreaksCustom}
\title{Custom minor breaks based on limits.}
\usage{
minorBreaksCustom(limits, breaksInterval)
}
\arguments{
\item{limits}{limits, vector of 2 POSIXct values}

\item{ii}{minor breaks interval, in hours}
}
\description{
Custom minor breaks based on limits.
}
