% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_display_options.R
\name{minValueInPlot}
\alias{minValueInPlot}
\title{Automatically get the min value from the plot.}
\usage{
minValueInPlot(plot, variable)
}
\arguments{
\item{plot}{ggplot2 plot}

\item{variable}{variable of interest}
}
\value{
min value in all data layers
}
\description{
Automatically get the min value from the plot.
}
