% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cache.R
\docType{class}
\name{model_cache-class}
\alias{model_cache-class}
\title{Model cache class.}
\description{
Model cache class.
}
\section{Slots}{

\describe{
\item{\code{mod}}{rxode2 or mrgsolve model (in cache)}

\item{\code{eta_names}}{ETA names in model}

\item{\code{variable}}{variable name representing the predictions in model}
}}

