% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_observations_table_editor.R
\name{ObservationsTableEditor}
\alias{ObservationsTableEditor}
\title{Observations table editor.}
\usage{
ObservationsTableEditor(
  tableReact = NULL,
  ns = shiny::NS("observations_table"),
  fun = NULL,
  initialObs = 100,
  initialTime = "12:00"
)
}
\arguments{
\item{tableReact}{reactive table}

\item{ns}{shiny namespace}

\item{fun}{conversion function (table -> anything)}

\item{initialObs}{first observation in table, only if tableReact is NULL}

\item{initialTime}{first observation time in table, only if tableReact is NULL}
}
\description{
Observations table editor.
}
