% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_options.R
\docType{class}
\name{plot_options-class}
\alias{plot_options-class}
\title{Plot options class.}
\description{
Plot options class.
}
\section{Slots}{

\describe{
\item{\code{ylim}}{suggested Y-axis limit}

\item{\code{timeref}}{POSIXct time corresponding to TIME=0}

\item{\code{date_labels}}{label of datetime on X-axis}

\item{\code{date_breaks}}{data breaks, e.g. '1 day'}

\item{\code{minor_breaks_interval}}{minor breaks interval in hours}
}}

