% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_display_options.R
\name{PlotDisplayOptions}
\alias{PlotDisplayOptions}
\title{Plot display options.}
\usage{
PlotDisplayOptions(
  ylim = NULL,
  timeref = NULL,
  date_labels = "\%b \%d",
  date_breaks = "1 day",
  minor_breaks_interval = 6,
  date_limits = .POSIXct(character(0)),
  show_legend = FALSE,
  legend_title = "Legend",
  legend_position = "right",
  x_axis_label = "Time",
  y_axis_label = "Concentration",
  x_axis_bar_plot_label = "Time",
  y_axis_bar_plot_label = "Dose"
)
}
\arguments{
\item{ylim}{suggested Y-axis limit}

\item{timeref}{POSIXct time corresponding to TIME=0 in simulation}

\item{date_labels}{label of datetime on X-axis}

\item{date_breaks}{data breaks, e.g. '1 day'}

\item{minor_breaks_interval}{minor breaks interval in hours}

\item{date_limits}{date limits, vector of 2 POSIXct values with the limits}

\item{show_legend}{show a legend}

\item{legend_title}{legend title}

\item{legend_position}{legend position}

\item{x_axis_label}{label}

\item{y_axis_label}{label}

\item{x_axis_bar_plot_label}{label}

\item{y_axis_bar_plot_label}{label}
}
\value{
an object
}
\description{
Plot display options.
}
