% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_options.R
\name{PlotOptions}
\alias{PlotOptions}
\title{Plot options.}
\usage{
PlotOptions(
  ylim = NULL,
  timeref = NULL,
  date_labels = "\%b \%d",
  date_breaks = "1 day",
  minor_breaks_interval = 6
)
}
\arguments{
\item{ylim}{suggested Y-axis limit}

\item{timeref}{POSIXct time corresponding to TIME=0 in simulation}

\item{date_labels}{label of datetime on X-axis}

\item{date_breaks}{data breaks, e.g. '1 day'}

\item{minor_breaks_interval}{minor breaks interval in hours}
}
\value{
an object
}
\description{
Plot options.
}
