% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_display_options.R
\name{plotToPOSIXct}
\alias{plotToPOSIXct}
\title{Automatically convert all TIME columns found in plot to POSIXct given the time reference.}
\usage{
plotToPOSIXct(plot, timeref)
}
\arguments{
\item{plot}{ggplot2 plot}

\item{timeref}{time reference}
}
\value{
updated plot
}
\description{
Automatically convert all TIME columns found in plot to POSIXct given the time reference.
}
