% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_formula.R
\name{populationLikelihood}
\alias{populationLikelihood}
\title{Population likelihood (how far the individual parameters deviate from 0).}
\usage{
populationLikelihood(model, etas)
}
\arguments{
\item{model}{Campsismap model}

\item{etas}{numeric vector}
}
\value{
population likelihood
}
\description{
Population likelihood (how far the individual parameters deviate from 0).
}
