% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_datetime_table_editor.R
\name{posixToDateStr}
\alias{posixToDateStr}
\title{Posix to date string.}
\usage{
posixToDateStr(x)
}
\arguments{
\item{x}{POSIXct date}
}
\value{
date, string form '%Y-%m-%d'
}
\description{
Posix to date string.
}
