% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_datetime_table_editor.R
\name{posixToTimeStr}
\alias{posixToTimeStr}
\title{Posix to time string.}
\usage{
posixToTimeStr(x, seconds = FALSE)
}
\arguments{
\item{x}{POSIXct time}

\item{seconds}{export seconds, logical value}
}
\value{
time, string form '%H:%M:%S' or '%H:%M' depending on argument 'seconds'
}
\description{
Posix to time string.
}
