% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/model_cache.R,
%   R/campsismap_model.R
\name{predict}
\alias{predict}
\alias{predict,rxode2_model_cache,tbl_df,numeric,simulation_settings-method}
\alias{predict,mrgsolve_model_cache,tbl_df,numeric,simulation_settings-method}
\alias{predict,campsismap_model,dataset,numeric,simulation_settings-method}
\title{Predict.}
\usage{
predict(object, dataset, etas = NULL, settings = NULL, ...)

\S4method{predict}{rxode2_model_cache,tbl_df,numeric,simulation_settings}(object, dataset, etas = NULL, settings = NULL, ...)

\S4method{predict}{mrgsolve_model_cache,tbl_df,numeric,simulation_settings}(object, dataset, etas = NULL, settings = NULL, ...)

\S4method{predict}{campsismap_model,dataset,numeric,simulation_settings}(object, dataset, etas = NULL, settings = NULL, ...)
}
\arguments{
\item{object}{campsismap object or model cache}

\item{dataset}{dataset}

\item{etas}{individual parameters to simulate}

\item{settings}{simulation settings}

\item{...}{extra arguments}
}
\value{
updated model
}
\description{
Predict.
}
