% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.R
\docType{class}
\name{proportional_error_model-class}
\alias{proportional_error_model-class}
\title{Proportional error model class.}
\description{
Proportional error model class.
}
\section{Slots}{

\describe{
\item{\code{prop}}{standard deviation of proportional error, numeric}
}}

