% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.R
\name{ProportionalErrorModel}
\alias{ProportionalErrorModel}
\title{Proportional error model.}
\usage{
ProportionalErrorModel(prop)
}
\arguments{
\item{prop}{standard deviation of proportional error, numeric}
}
\value{
error model
}
\description{
Proportional error model.
}
