% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_datetime_table_editor.R
\name{read.datetimecsv}
\alias{read.datetimecsv}
\title{Read 'Datetime' csv file.
This csv file must contain at least a column 'Date' and a column 'Time'.
Date column must be formatted as 'ymd'.
Time column must be formatted as 'hm'.}
\usage{
read.datetimecsv(
  file,
  dateTime = TRUE,
  sort = TRUE,
  relative = FALSE,
  dateTime0 = NULL
)
}
\arguments{
\item{file}{path to csv file}

\item{dateTime}{if TRUE, the 'Date' and 'Time' columns will be converted into 'Datetime'}

\item{sort}{by default, rows will be sorted (only if datetime is TRUE)}

\item{relative}{Datetime replaced by relative TIME column (based on provided dateTime0)}

\item{dateTime0}{reference date if relative argument is used, otherwise first datetime found in table is used as reference datetime}
}
\description{
Read 'Datetime' csv file.
This csv file must contain at least a column 'Date' and a column 'Time'.
Date column must be formatted as 'ymd'.
Time column must be formatted as 'hm'.
}
