% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_inputs.R
\name{read.inputs}
\alias{read.inputs}
\title{Read e-Campsis inputs.}
\usage{
read.inputs(content, defaultInputs)
}
\arguments{
\item{content}{object returned by method fromJSON}

\item{defaultInputs}{default inputs that must be read}
}
\value{
e-Campsis inputs
}
\description{
Read e-Campsis inputs.
}
