% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/recommend.R
\name{recommend}
\alias{recommend}
\alias{recommend,campsismap_model,dataset,numeric,target_definition,POSIXct,simulation_settings-method}
\title{Recommend.}
\usage{
recommend(
  object,
  dataset,
  etas = NULL,
  target = NULL,
  now = NULL,
  settings = NULL,
  ...
)

\S4method{recommend}{campsismap_model,dataset,numeric,target_definition,POSIXct,simulation_settings}(
  object,
  dataset,
  etas = NULL,
  target = NULL,
  now = NULL,
  settings = NULL,
  ...
)
}
\arguments{
\item{object}{campsismap object or model cache}

\item{dataset}{dataset}

\item{etas}{individual parameters to simulate}

\item{target}{target definition}

\item{now}{defines what the future or the past is, POSIXct date}

\item{settings}{simulation settings}

\item{...}{extra arguments}
}
\value{
recommendations
}
\description{
Recommend.
}
