% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendation.R
\name{Recommendation}
\alias{Recommendation}
\title{Make a recommendation object.}
\usage{
Recommendation(dataset, target, rules, now = 0)
}
\arguments{
\item{dataset}{original dataset}

\item{target}{target definition}

\item{rules}{dose adaptation rules}

\item{now}{numeric value}
}
\value{
recommendation object
}
\description{
Make a recommendation object.
}
