% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plot_type.R
\docType{class}
\name{recommendation_full_plot_type-class}
\alias{recommendation_full_plot_type-class}
\title{Recommendation full plot type (profile + bar plot).}
\description{
Recommendation full plot type (profile + bar plot).
}
