% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plot_type.R
\name{RecommendationBarPlotType}
\alias{RecommendationBarPlotType}
\title{Recommendation bar plot type.}
\usage{
RecommendationBarPlotType()
}
\value{
recommendation bar plot type
}
\description{
Recommendation bar plot type.
}
