% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plot_type.R
\name{RecommendationFullPlotType}
\alias{RecommendationFullPlotType}
\title{Recommendation full plot type.}
\usage{
RecommendationFullPlotType()
}
\value{
recommendation full plot type
}
\description{
Recommendation full plot type.
}
