% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plot_type.R
\name{RecommendationPlotType}
\alias{RecommendationPlotType}
\title{Recommendation plot type.}
\usage{
RecommendationPlotType()
}
\value{
recommendation plot type
}
\description{
Recommendation plot type.
}
