% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_inputs.R
\name{replaceInput}
\alias{replaceInput}
\alias{replaceInput,shiny_inputs,shiny_input-method}
\title{Replace input.}
\usage{
replaceInput(object, x, newvalue)

\S4method{replaceInput}{shiny_inputs,shiny_input}(object, x, newvalue)
}
\arguments{
\item{object}{generic object}

\item{x}{input to replace}

\item{newvalue}{input updated value}
}
\value{
updated object
}
\description{
Replace input.
}
