% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/dataset.R, R/recommendation.R
\name{retrieveDoseAmount}
\alias{retrieveDoseAmount}
\alias{retrieveDoseAmount,dataset,integer-method}
\alias{retrieveDoseAmount,recommendation,integer-method}
\title{Retrieve dose amount.}
\usage{
retrieveDoseAmount(object, dose_number, ...)

\S4method{retrieveDoseAmount}{dataset,integer}(object, dose_number)

\S4method{retrieveDoseAmount}{recommendation,integer}(object, dose_number)
}
\arguments{
\item{object}{generic object (e.g. dataset)}

\item{dose_number}{corresponding dose number}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Retrieve dose amount.
}
