% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_generic.R,
%   R/shiny_datetime_table_editor.R
\name{server}
\alias{server}
\alias{server,datetime_table_editor-method}
\title{Call to server.}
\usage{
server(object, input, output, session, ...)

\S4method{server}{datetime_table_editor}(object, input, output, session, dateTime0React, nowReact = NULL)
}
\arguments{
\item{object}{generic object}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{...}{extra arguments}

\item{dateTime0React}{reference datetime}

\item{nowReact}{now datetime}
}
\value{
this is up to the implementation
}
\description{
Call to server.
}
