% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_generic.R
\name{serverFunction}
\alias{serverFunction}
\title{Get shiny server function.}
\usage{
serverFunction(object)
}
\arguments{
\item{object}{generic object}
}
\value{
a server function for shiny
}
\description{
Get shiny server function.
}
