% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/campsismap_model.R
\name{setup}
\alias{setup}
\alias{setup,campsismap_model,character,simulation_settings-method}
\title{Setup campsismap object by caching the model.}
\usage{
setup(object, dest, settings = NULL, ...)

\S4method{setup}{campsismap_model,character,simulation_settings}(object, dest, settings)
}
\arguments{
\item{object}{campsismap object}

\item{dest}{destination engine}

\item{settings}{simulation settings}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Setup campsismap object by caching the model.
}
