% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/model_cache.R
\name{setupModel}
\alias{setupModel}
\alias{setupModel,rxode2_model_cache,simulation_settings-method}
\alias{setupModel,mrgsolve_model_cache,simulation_settings-method}
\title{Setup model (once before the estimation and after the model compilation).}
\usage{
setupModel(object, settings, ...)

\S4method{setupModel}{rxode2_model_cache,simulation_settings}(object, settings, ...)

\S4method{setupModel}{mrgsolve_model_cache,simulation_settings}(object, settings, ...)
}
\arguments{
\item{object}{model cache}

\item{settings}{simulation settings}

\item{...}{extra arguments}
}
\value{
updated model
}
\description{
Setup model (once before the estimation and after the model compilation).
}
