% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_input.R
\name{shinyInput}
\alias{shinyInput}
\title{Create a new shiny input.}
\usage{
shinyInput(id, value, transient = FALSE)
}
\arguments{
\item{id}{input identifier}

\item{value}{input default value}

\item{transient}{transient input (i.e. not persisted)}
}
\value{
a shiny input
}
\description{
Create a new shiny input.
}
