% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_definition.R
\docType{class}
\name{target_definition_effective-class}
\alias{target_definition_effective-class}
\title{Target definition (effective).
TIME   VALUE LAST_DOSENO LAST_DOSE_TIME 
24     X     1            0               Predose 2 should attain a target of X
48     Y     2            24              Predose 3 should attain a target of Y}
\description{
Target definition (effective).
TIME   VALUE LAST_DOSENO LAST_DOSE_TIME 
24     X     1            0               Predose 2 should attain a target of X
48     Y     2            24              Predose 3 should attain a target of Y
}
\section{Slots}{

\describe{
\item{\code{table}}{table}
}}

