% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_definition.R
\docType{class}
\name{target_definition_per_dose-class}
\alias{target_definition_per_dose-class}
\title{Target definition per dose.
DOSENO VALUE
1      X   Dose 1 should attain a target of X
2      Y   Dose 2 should attain a target of Y, etc.}
\description{
Target definition per dose.
DOSENO VALUE
1      X   Dose 1 should attain a target of X
2      Y   Dose 2 should attain a target of Y, etc.
}
\section{Slots}{

\describe{
\item{\code{table}}{table}
}}

