% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_definition.R
\docType{class}
\name{target_definition_per_window-class}
\alias{target_definition_per_window-class}
\title{Target definition per window.}
\description{
Target definition per window.
}
\section{Slots}{

\describe{
\item{\code{table}}{table}
}}

