% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_definition.R
\name{TargetDefinitionEffective}
\alias{TargetDefinitionEffective}
\title{Target definition (effective).}
\usage{
TargetDefinitionEffective(table)
}
\arguments{
\item{table}{data frame with (effective) TIME and VALUE columns}
}
\description{
Target definition (effective).
}
