% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_definition.R
\name{TargetDefinitionPerDose}
\alias{TargetDefinitionPerDose}
\title{Target definition per dose.}
\usage{
TargetDefinitionPerDose(table)
}
\arguments{
\item{table}{data frame with DOSENO and VALUE columns}
}
\description{
Target definition per dose.
}
