% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_definition.R
\name{TargetDefinitionPerWindow}
\alias{TargetDefinitionPerWindow}
\title{Target definition per window.
TIME VALUE
0    X   Target of X from 0 to 24 (0 or negative value must be provided)
24   Y   Target of Y from 24 to Inf, etc.}
\usage{
TargetDefinitionPerWindow(table)
}
\arguments{
\item{table}{data frame with TIME and VALUE columns}
}
\description{
Target definition per window.
TIME VALUE
0    X   Target of X from 0 to 24 (0 or negative value must be provided)
24   Y   Target of Y from 24 to Inf, etc.
}
