% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_datetime_table_editor.R
\name{toDateTime}
\alias{toDateTime}
\title{To date time.}
\usage{
toDateTime(date, time)
}
\arguments{
\item{date}{date, 'ymd' formatted string}

\item{time}{time, 'hm' formatted string}
}
\value{
POSIXct value
}
\description{
To date time.
}
