% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_datetime_table_editor.R
\name{toRelativeTimeTable}
\alias{toRelativeTimeTable}
\title{Add relative time column to the datetime table.}
\usage{
toRelativeTimeTable(table, dateTime0)
}
\arguments{
\item{table}{table content (see slot tableReact)}

\item{dateTime0}{reference datetime}
}
\value{
table with relative time column (column 'TIME')
}
\description{
Add relative time column to the datetime table.
}
