% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/dataset.R
\name{updateDoseAmount}
\alias{updateDoseAmount}
\alias{updateDoseAmount,dataset,numeric,integer-method}
\title{Update dose amount.}
\usage{
updateDoseAmount(object, amount, dose_number, ...)

\S4method{updateDoseAmount}{dataset,numeric,integer}(object, amount, dose_number)
}
\arguments{
\item{object}{generic object (e.g. dataset)}

\item{amount}{updated dose amount}

\item{dose_number}{corresponding dose number}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Update dose amount.
}
