% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_inputs.R
\name{updateInputs}
\alias{updateInputs}
\alias{updateInputs,shiny_inputs-method}
\title{Update inputs from shiny GUI.}
\usage{
updateInputs(object, input)

\S4method{updateInputs}{shiny_inputs}(object, input)
}
\arguments{
\item{object}{shiny inputs (ecampsis object)}

\item{input}{shiny input list from GUI}
}
\value{
updated inputs
}
\description{
Update inputs from shiny GUI.
}
