% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsismap_model.R
\docType{class}
\name{campsismap_model-class}
\alias{campsismap_model-class}
\title{Campsismap model class.}
\description{
Campsismap model class.
}
\section{Slots}{

\describe{
\item{\code{model}}{Campsis model ready for estimation}

\item{\code{omega}}{pre-computed OMEGA matrix}

\item{\code{eta_names}}{ETA names in model}

\item{\code{variable}}{variable name representing the predictions in model}

\item{\code{error}}{error model}

\item{\code{model_cache}}{model cache}

\item{\code{settings}}{simulation settings}

\item{\code{dest}}{destination engine}
}}

