% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsismap_model.R
\name{CampsismapModel}
\alias{CampsismapModel}
\title{Create a new Campsismap model.}
\usage{
CampsismapModel(model, variable)
}
\arguments{
\item{model}{Campsis model}

\item{variable}{variable of the concentration in model}
}
\value{
a Campsismap mode
}
\description{
Create a new Campsismap model.
}
