% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_datetime_table_editor.R,
%   R/shiny_dosing_table_editor.R, R/shiny_observations_table_editor.R,
%   R/shiny_target_table_editor.R
\name{getInitialTable}
\alias{getInitialTable}
\alias{getInitialTable,dosing_table_editor-method}
\alias{getInitialTable,observations_table_editor-method}
\alias{getInitialTable,target_table_editor-method}
\title{Get initial table.}
\usage{
getInitialTable(object, ...)

\S4method{getInitialTable}{dosing_table_editor}(object, init_dose, init_time)

\S4method{getInitialTable}{observations_table_editor}(object)

\S4method{getInitialTable}{target_table_editor}(object, init_target, init_time)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments}

\item{init_dose}{initial dose in table}

\item{init_time}{initial time in table}

\item{init_target}{initial dose in table}
}
\value{
initial table, tibble
}
\description{
Get initial table.
}
