% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_formula.R
\name{individualLikelihood}
\alias{individualLikelihood}
\title{Individual likelihood (how far the individual concentrations deviate from the observations, weighted by the error model).}
\usage{
individualLikelihood(model, dataset, samples, etas)
}
\arguments{
\item{model}{Campsismap model}

\item{dataset}{exported dataset}

\item{samples}{individual samples}

\item{etas}{unnamed numeric vector}
}
\value{
population likelihood
}
\description{
Individual likelihood (how far the individual concentrations deviate from the observations, weighted by the error model).
}
