% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/shiny_datetime_table_editor.R
\name{load}
\alias{load}
\alias{load,datetime_table_editor,character-method}
\title{Generic load method to load data from a file or a folder.}
\usage{
load(object, file, ...)

\S4method{load}{datetime_table_editor,character}(object, file, ...)
}
\arguments{
\item{object}{datetime table editor}

\item{file}{path to csv file that contains 'Date' and 'Time' information}

\item{...}{extra arguments, unused}
}
\value{
the loaded object

TRUE
}
\description{
Generic load method to load data from a file or a folder.

Read 'Datetime' csv file and load the editor.
See method 'read.datetimecsv'.
}
