% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/shiny_datetime_table_editor.R
\name{load}
\alias{load}
\alias{read.datetimecsv}
\title{Generic load method to load data from a file or a folder.}
\usage{
load(object, file, ...)

read.datetimecsv(
  file,
  dateTime = TRUE,
  sort = TRUE,
  relative = FALSE,
  dateTime0 = NULL
)
}
\arguments{
\item{object}{object to load with something}

\item{file}{path to csv file}

\item{...}{extra arguments}

\item{dateTime}{if TRUE, the 'Date' and 'Time' columns will be converted into 'Datetime'}

\item{sort}{by default, rows will be sorted (only if datetime is TRUE)}

\item{relative}{Datetime replaced by relative TIME column (based on provided dateTime0)}

\item{dateTime0}{reference date if relative argument is used, otherwise first datetime found in table is used as reference datetime}
}
\value{
the loaded object
}
\description{
Generic load method to load data from a file or a folder.

Read 'Datetime' csv file.
This csv file must contain at least a column 'Date' and a column 'Time'.
Date column must be formatted as 'ymd'.
Time column must be formatted as 'hm'.
}
