% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/shiny_datetime_table_editor.R
\name{load}
\alias{load}
\alias{read.datetimecsv}
\title{Generic load method to load data from a file or a folder.}
\usage{
load(object, file, ...)

read.datetimecsv(file)
}
\arguments{
\item{object}{object to load with something}

\item{file}{path to csv file}

\item{...}{extra arguments}
}
\value{
the loaded object
}
\description{
Generic load method to load data from a file or a folder.

Read 'Datetime' csv file.
This csv file must contain at least a column 'Date' and a column 'Time'.
}
