% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{load}
\alias{load}
\title{Generic load method to load data from a file or a folder.}
\usage{
load(object, file, ...)
}
\arguments{
\item{object}{object to load with something}

\item{file}{path to the file or folder to be read}

\item{...}{extra arguments}
}
\value{
the loaded object
}
\description{
Generic load method to load data from a file or a folder.
}
