% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_observations_table_editor.R
\name{ObservationsTableEditor}
\alias{ObservationsTableEditor}
\title{Observations table editor.}
\usage{
ObservationsTableEditor(
  tableReact = NULL,
  ns = shiny::NS("observations_table"),
  fun = NULL,
  defaultObs = 100,
  defaultTime = "06:00",
  greyOutPast = FALSE
)
}
\arguments{
\item{tableReact}{reactive table}

\item{ns}{shiny namespace}

\item{fun}{conversion function (table -> anything)}

\item{defaultObs}{first observation in table, only if tableReact is NULL}

\item{defaultTime}{first observation time in table, only if tableReact is NULL}

\item{greyOutPast}{grey out the rows in past, logical value}
}
\description{
Observations table editor.
}
