% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_display_options.R
\docType{class}
\name{plot_display_options-class}
\alias{plot_display_options-class}
\title{Plot options class.}
\description{
Plot options class.
}
\section{Slots}{

\describe{
\item{\code{ylim}}{suggested Y-axis limit}

\item{\code{ylim_bar_plot}}{suggested Y-axis limit in bar plot}

\item{\code{timeref}}{POSIXct time corresponding to TIME=0}

\item{\code{date_labels}}{label of datetime on X-axis}

\item{\code{date_breaks}}{data breaks, e.g. '1 day'}

\item{\code{minor_breaks_interval}}{minor breaks interval in hours}

\item{\code{date_limits}}{date limits, vector of 2 POSIXct values with the limits}

\item{\code{show_legend}}{show a legend}

\item{\code{legend_title}}{legend title}

\item{\code{legend_position}}{legend position}

\item{\code{x_axis_label}}{label}

\item{\code{y_axis_label}}{label}

\item{\code{x_axis_bar_plot_label}}{label}

\item{\code{y_axis_bar_plot_label}}{label}

\item{\code{bar_plot_value_mode}}{bar plot value mode: 'within' or 'above'}
}}

