% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_display_options.R
\name{PlotDisplayOptions}
\alias{PlotDisplayOptions}
\title{Plot display options.}
\usage{
PlotDisplayOptions(
  ylim = NULL,
  timeref = NULL,
  date_labels = "\%b \%d",
  date_breaks = "1 day",
  minor_breaks_interval = 6,
  show_legend = FALSE,
  legend_title = "Legend"
)
}
\arguments{
\item{ylim}{suggested Y-axis limit}

\item{timeref}{POSIXct time corresponding to TIME=0 in simulation}

\item{date_labels}{label of datetime on X-axis}

\item{date_breaks}{data breaks, e.g. '1 day'}

\item{minor_breaks_interval}{minor breaks interval in hours}

\item{show_legend}{show a legend}

\item{legend_title}{legend title}
}
\value{
an object
}
\description{
Plot display options.
}
