% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_display_options.R
\name{PlotDisplayOptions}
\alias{PlotDisplayOptions}
\title{Plot display options.}
\usage{
PlotDisplayOptions(
  ylim = NULL,
  ylim_bar_plot = NULL,
  timeref = NULL,
  date_labels = "\%b \%d",
  date_breaks = "1 day",
  minor_breaks_interval = 6,
  date_limits = .POSIXct(character(0)),
  date_n_dodge = 1L,
  show_legend = FALSE,
  legend_title = "Legend",
  legend_position = "right",
  x_axis_label = "Time",
  y_axis_label = "Concentration",
  x_axis_bar_plot_label = "Time",
  y_axis_bar_plot_label = "Dose",
  bar_plot_value_mode = "within",
  typical_profile_colour = "#6196B4",
  individual_fit_profile_colour = "#B90E1E",
  recommendation_profile_colour = "#B2B2B2",
  individual_fit_bar_colour = "#B90E1E",
  recommendation_bar_colour = "#B2B2B2",
  target_profile_colour = "palegreen1"
)
}
\arguments{
\item{ylim}{suggested Y-axis limit}

\item{ylim_bar_plot}{suggested Y-axis limit in bar plot}

\item{timeref}{POSIXct time corresponding to TIME=0 in simulation}

\item{date_labels}{label of datetime on X-axis}

\item{date_breaks}{data breaks, e.g. '1 day'}

\item{minor_breaks_interval}{minor breaks interval in hours}

\item{date_limits}{date limits, vector of 2 POSIXct values with the limits}

\item{date_n_dodge}{number of dodges for date}

\item{show_legend}{show a legend}

\item{legend_title}{legend title}

\item{legend_position}{legend position}

\item{x_axis_label}{label}

\item{y_axis_label}{label}

\item{x_axis_bar_plot_label}{label}

\item{y_axis_bar_plot_label}{label}

\item{bar_plot_value_mode}{bar plot value mode: 'within' or 'above'}

\item{typical_profile_colour}{colour of typical profile}

\item{individual_fit_profile_colour}{colour of individual fit profile}

\item{recommendation_profile_colour}{colour of recommendation profile}

\item{individual_fit_bar_colour}{colour of individual fit bar}

\item{recommendation_bar_colour}{colour of recommendation bar}

\item{target_profile_colour}{colour of target profile}
}
\value{
an object
}
\description{
Plot display options.
}
