% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_options.R
\name{PlotOptions}
\alias{PlotOptions}
\title{Plot options.}
\usage{
PlotOptions(ylim = NULL, timeref = NULL, timelabel = "\%b \%d\\n\%H:\%M")
}
\arguments{
\item{ylim}{suggested Y-axis limit}

\item{timeref}{POSIXct time corresponding to TIME=0 in simulation}

\item{timelabel}{label of datetime on X-axis}
}
\value{
an object
}
\description{
Plot options.
}
