% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_model.R
\name{ProportionalErrorModel}
\alias{ProportionalErrorModel}
\title{Proportional error model.}
\usage{
ProportionalErrorModel(prop)
}
\description{
Proportional error model.
}
\section{Slots}{

\describe{
\item{\code{prop}}{standard deviation of proportional error, numeric}
}}

