% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/quick_plot.R
\name{quickPlot}
\alias{quickPlot}
\alias{quickPlot,campsismap_model,dataset,numeric,individual_fit_plot_type,plot_display_options-method}
\alias{quickPlot,campsismap_model,dataset,numeric,recommendation_plot_type,plot_display_options-method}
\title{Quickly visualise the individual predictions with regard to the observed data.}
\usage{
quickPlot(model, dataset = NULL, etas = NULL, plot = NULL, options = NULL, ...)

\S4method{quickPlot}{campsismap_model,dataset,numeric,individual_fit_plot_type,plot_display_options}(model, dataset, etas, plot, options)

\S4method{quickPlot}{campsismap_model,dataset,numeric,recommendation_plot_type,plot_display_options}(model, dataset, etas, plot, options, recommendation)
}
\arguments{
\item{model}{Campsismap or Campsis model}

\item{dataset}{Campsis dataset}

\item{etas}{individual parameters}

\item{plot}{plot type}

\item{options}{plot options}

\item{...}{extra arguments}
}
\value{
a ggplot2 object
}
\description{
Quickly visualise the individual predictions with regard to the observed data.
}
