% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_definition.R
\docType{class}
\name{target_definition_effective-class}
\alias{target_definition_effective-class}
\title{Target definition (effective).
TIME   VALUE
24     X   Predose 2 should attain a target of X
48     Y   Predose 3 should attain a target of Y}
\description{
Target definition (effective).
TIME   VALUE
24     X   Predose 2 should attain a target of X
48     Y   Predose 3 should attain a target of Y
}
\section{Slots}{

\describe{
\item{\code{table}}{table}
}}

