% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_definition.R
\name{TargetDefinitionEffective}
\alias{TargetDefinitionEffective}
\title{Target definition (effective).}
\usage{
TargetDefinitionEffective(
  table = data.frame(TIME = numeric(0), VALUE = numeric(0), LAST_DOSENO = integer(0),
    LAST_DOSE_TIME = numeric(0))
)
}
\arguments{
\item{table}{data frame with (effective) TIME and VALUE columns}
}
\description{
Target definition (effective).
}
