% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trough_time_rule.R
\name{TroughTimeRule}
\alias{TroughTimeRule}
\title{Rule to identity when the trough time is.}
\usage{
TroughTimeRule(ii = 24, use_next_dose = TRUE)
}
\arguments{
\item{ii}{dosing interval}

\item{use_next_dose}{if TRUE next dose will be used as trough time, if FALSE dose + ii is used as trough time}
}
\description{
Rule to identity when the trough time is.
}
